refactor ticket list view with styles
